components: migrate YearsColumn to typescript
